Moved DecoratorDiverted annotation to package "com.github.cloudyrock.mongock.internal".
