+++ mongock-api/src/main/java/com/github/cloudyrock/mongock/internal/DecoratorDiverted.java
@@ -1,4 +1,4 @@
-package io.changock.migration.api.annotations;
+package com.github.cloudyrock.mongock.internal;
 
 import com.github.cloudyrock.mongock.ChangeLog;
