+++ mongock-decorator-util/src/main/java/com/github/cloudyrock/mongock/test/util/decorator/DecoratorValidator.java
@@ -4,7 +4,7 @@ import com.github.cloudyrock.mongock.driver.api.lock.LockManager;
 import com.github.cloudyrock.mongock.driver.api.lock.guard.decorator.DecoratorBase;
 import com.github.cloudyrock.mongock.driver.api.lock.guard.invoker.LockGuardInvoker;
 import com.github.cloudyrock.mongock.driver.api.lock.guard.invoker.LockGuardInvokerImpl;
-import io.changock.migration.api.annotations.DecoratorDiverted;
+import com.github.cloudyrock.mongock.internal.DecoratorDiverted;
 import io.changock.migration.api.annotations.NonLockGuarded;
 import io.changock.migration.api.annotations.NonLockGuardedType;
 import org.mockito.Mockito;
